test: consolidate identical StubHandler into shared TestSupport (#144)
